refactor: move tests to their respective directories
